refactor: extract DropdownButton
