+++ src/widgets/ArticlesFilter/styles/cat_filter_button.ts
@@ -1,9 +1,6 @@
 import styled from 'styled-components'
 
-import Button from '@/widgets/Buttons/Button'
 import css, { theme } from '@/utils/css'
-
-import ArrowSVG from '@/icons/ArrowSimple'
 
 export const Wrapper = styled.div`
   ${css.flex('align-center')};
@@ -13,30 +10,4 @@ export const Wrapper = styled.div`
 `
 export const Label = styled.div`
   opacity: 0.7;
-`
-export const InnerBtnWrapper = styled.div`
-  ${css.flex('align-center')};
-  margin-left: 2px;
-  color: ${theme('thread.articleDigest')};
-  font-weight: 400;
-  font-size: 13px;
-
-  &:hover {
-    cursor: pointer;
-    color: ${theme('thread.articleTitle')};
-  }
-
-  transition: color 0.2s;
-`
-export const ButtonWrapper = styled(Button)`
-  border: none;
-`
-export const FilterIcon = styled(ArrowSVG)`
-  fill: ${theme('thread.articleDigest')};
-  ${css.size(14)};
-  transform: rotate(-90deg);
-  margin-left: 5px;
-  ${InnerBtnWrapper}:hover & {
-    fill: ${theme('thread.articleDigest')};
-  }
 `
