+++ src/widgets/Buttons/PublishButton/PostLayout.tsx
@@ -9,7 +9,6 @@ import { memo, FC, Fragment } from 'react'
 import { buildLog } from '@/utils/logger'
 import { SpaceGrow } from '@/widgets/Common'
 
-// import DropdownButton from './DropdownButton'
 // import Button from './Button'
 import { Wrapper, Title, EditIcon } from '../styles/publish_button/post_layout'
