+++ src/widgets/Buttons/PublishButton/WorksLayout.tsx
@@ -7,9 +7,7 @@
 import { memo, FC } from 'react'
 
 import { buildLog } from '@/utils/logger'
-import { SpaceGrow } from '@/widgets/Common'
 
-// import DropdownButton from './DropdownButton'
 // import Button from './Button'
 import {
   Wrapper,
