+++ src/widgets/Buttons/styles/dropdown_button/index.ts
@@ -1,40 +1,48 @@
 import styled from 'styled-components'
 
-import Img from '@/Img'
-import { Wrapper as NormalButton } from '../button'
+import type { TSpace } from '@/spec'
+import Button from '@/widgets/Buttons/Button'
 import css, { theme } from '@/utils/css'
 
-export const Wrapper = styled.div`
-  ${css.flex('align-center')};
-  width: 100%;
-  position: relative;
-`
-export const ButtonWrapper = styled(NormalButton)`
-  width: 100%;
-  height: 30px;
-  border-top-right-radius: 0;
-  border-bottom-right-radius: 0;
-  padding-right: 6px;
-`
-export const DropdownButtonWrapper = styled(NormalButton)`
-  width: 36px;
-  height: 30px;
-  border-top-left-radius: 0;
-  border-bottom-left-radius: 0;
-`
-export const Divider = styled.div`
-  position: absolute;
-  right: 35px;
-  width: 1px;
-  height: 24px;
-  background: #2c6b94;
+import ArrowSVG from '@/icons/ArrowSimple'
 
-  ${Wrapper}:hover & {
-    display: none;
-  }
+export const Wrapper = styled.div<TSpace>`
+  ${css.flex('align-center')};
+  color: ${theme('thread.articleDigest')};
+  margin-right: -12px;
+  font-size: 13px;
+
+  margin-top: ${({ top }) => `${top}px`};
+  margin-bottom: ${({ bottom }) => `${bottom}px`};
+  margin-left: ${({ left }) => `${left}px`};
+  margin-right: ${({ right }) => `${right}px`};
 `
-export const MoreIcon = styled(Img)`
-  fill: ${theme('button.fg')};
-  ${css.size(10)};
-  transform: rotate(180deg);
+export const Label = styled.div`
+  opacity: 0.7;
+`
+export const InnerBtnWrapper = styled.div`
+  ${css.flex('align-center')};
+  margin-left: 2px;
+  color: ${theme('thread.articleDigest')};
+  font-weight: 400;
+  font-size: 13px;
+
+  &:hover {
+    cursor: pointer;
+    color: ${theme('thread.articleTitle')};
+  }
+
+  transition: color 0.2s;
+`
+export const ButtonWrapper = styled(Button)`
+  border: none;
+`
+export const FilterIcon = styled(ArrowSVG)`
+  fill: ${theme('thread.articleDigest')};
+  ${css.size(14)};
+  transform: rotate(-90deg);
+  margin-left: 5px;
+  ${InnerBtnWrapper}:hover & {
+    fill: ${theme('thread.articleDigest')};
+  }
 `
